move pos settings mapping to yosemite
